lambda/resource_cost_monitoring_lambda_deploy/requirements.txt
boto3==1.35.91
pyyaml==6.0.2
botocore==1.35.91